pyproject.toml
[project]
name = "crewai-snowmcp"
version = "0.1.0"
description = "CrewAI integration with Snowflake MCP Server"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "crewai[google-genai]>=1.7.0",
    "crewai-tools>=1.7.0",
    "python-dotenv>=1.0.0",
    "requests>=2.31.0",
]
